xorgproto fontconfig libX11 mesa xserver pixman putty libxcb xcb-proto xkeyboard-config EGL-Registry git update Jun 2 2019
xserver          commit 26fe29f4fa53cbb7d51892e2cf397c084093812f
libxcb           commit be1745c8eb00defcb31d336ccc142de056e92bd8
libxcb/xcb-proto commit dbf346502a31c3629b2c597711fb4b0ea6521de6
xkeyboard-config commit 644a0820d1018ddcc782f522cde9eb136a437383
libX11           commit 99a2cf1aa0b58391078d5d3edf0a7dab18c7745d
pixman           commit cb2ec4268fbde0df3b588ce5cbe2e43e04654520
mesa             commit 75590604a96e266c9b466db3250994333ce4dcc8
putty            commit 45be166be3ac000b13cbc5f9a0d858c690a8ae4c
EGL-Registry     commit 4744552d13f4475839d45c2eae7f745bac6ca204
xorgproto        commit 6bb0d500e5a7200cddee2d967efedac50a52aa6c

+++ X11/Xalloca.h
@@ -31,7 +31,7 @@ from The Open Group.
  * These macros are used to make fast, function-local memory allocations.
  * Their characteristics are as follows:
  *
- * void *ALLOCATE_LOCAL(int size)
+ * void *ALLOCATE_LOCAL(size_t size)
  *    Returns a pointer to size bytes of memory, or NULL if the allocation
  *    failed.  The memory must be freed with DEALLOCATE_LOCAL before the
  *    function that made the allocation returns.  You should not ask for
@@ -80,7 +80,7 @@ from The Open Group.
 #    ifndef alloca
 #      define alloca __builtin_alloca
 #    endif /* !alloca */
-#    define ALLOCATE_LOCAL(size) alloca((int)(size))
+#    define ALLOCATE_LOCAL(size) alloca((size_t)(size))
 #  else /* ! __GNUC__ */
 
 /*
@@ -98,7 +98,7 @@ from The Open Group.
 #          define __Xnullarg		/* as nothing */
            extern void *alloca(__Xnullarg);
 #        endif
-#        define ALLOCATE_LOCAL(size) alloca((int)(size))
+#        define ALLOCATE_LOCAL(size) alloca((size_t)(size))
 #      endif /* who does alloca */
 #  endif /* __GNUC__ */
